+++ spec/doctest_helper.rb
@@ -5,3 +5,4 @@ require_relative 'support/fake_db'
 require_relative 'support/fake_models'
 
 require "teckel"
+require "teckel/chain"