+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changes
 
-- Breaking: `Teckel::Chain` will not be required by default. require manually if needed `require "teckel/chain"`
-- Breaking: Internally, `Teckel::Operation::Runner` instead of `:success` and `:failure` now only uses `:halt` as it's throw-catch symbol.
+- Breaking: `Teckel::Chain` will not be required by default. require manually if needed `require "teckel/chain"` [GH-24]
+- Breaking: Internally, `Teckel::Operation::Runner` instead of `:success` and `:failure` now only uses `:halt` as it's throw-catch symbol. [GH-26]
 - Add: Using the default `Teckel::Operation::Runner`, `input_constructor` and `result_constructor` will be executed
   within the context of the operation instance. This allows for `input_constructor` to call `fail!` and `success!` 
-  without ever `call`ing the operation
+  without ever `call`ing the operation. [GH-26]
 
 
 ## 0.6.0